LFS-891: Re-purpose the global login context as a global app context rename to appContext
move out of loginDialog and to its own file in the commons module
